Add preferences.xhtml supplying a preference pane
Puts back in preferences.xhtml and preferences.js, and adds extensions.make-it-red-text-color as a Pref in prefs.js. This makes the Make It Red preference pane in Settings work again.
The .xhtml is close to the one that's still available in https://zotero-download.s3.amazonaws.com/tmp/make-it-red/make-it-red-2.0.xpi, however, here:
- there is no <h2> title, this is not needed since Zotero already supplies a title
- it uses <html:input> instead of <input>, this way the color picker works (at least in Windows 11, for me)
- it comes with proper l10n