Make the style.css color a variable
This makes the style.css color a variable, and initializes it from the corresponding Settings. As a result, the Make It Red text color settings in the Settings pane actually have an effect.
(We have to make the green color a variable too to get valid CSS that plays nice with the menu item View/Make It Green Instead.)